+++ examples/swr-site/next.config.js
@@ -65,8 +65,5 @@ module.exports = withNextra({
   reactStrictMode: true,
   experimental: {
     newNextLinkBehavior: true,
-    images: {
-      allowFutureImage: true,
-    },
   },
 });