+++ packages/nextra-theme-docs/src/components/theme-switch.tsx
@@ -8,35 +8,36 @@ type ThemeSwitchProps = {
   lite?: boolean
 }
 
+const OPTIONS = [
+  { key: 'light', name: 'Light' },
+  { key: 'dark', name: 'Dark' },
+  { key: 'system', name: 'System' }
+]
+
 export function ThemeSwitch({ lite }: ThemeSwitchProps): ReactElement {
-  const { theme, setTheme, systemTheme } = useTheme()
-  const renderedTheme = theme === 'system' ? systemTheme : theme
+  const { setTheme, resolvedTheme, theme = '' } = useTheme()
   const mounted = useMounted()
-  const IconToUse = mounted && renderedTheme === 'dark' ? MoonIcon : SunIcon
+  const IconToUse = mounted && resolvedTheme === 'dark' ? MoonIcon : SunIcon
   return (
-  <div className="relative">
-    <Select
-      title="Change theme"
-      onChange={option => {
-        setTheme(option.key)
-      }}
-      selected={{
-        key: theme || '',
-        name: (
-          <div className="flex items-center gap-2 capitalize">
-            <IconToUse />
-            <span className={lite ? 'md:hidden' : ''}>
-              {mounted ? theme : 'light'}
-            </span>
-          </div>
-        )
-      }}
-      options={[
-        { key: 'light', name: 'Light' },
-        { key: 'dark', name: 'Dark' },
-        { key: 'system', name: 'System' }
-      ]}
-    />
-  </div>
+    <div className="nx-relative">
+      <Select
+        title="Change theme"
+        options={OPTIONS}
+        onChange={option => {
+          setTheme(option.key)
+        }}
+        selected={{
+          key: theme,
+          name: (
+            <div className="nx-flex nx-items-center nx-gap-2 nx-capitalize">
+              <IconToUse />
+              <span className={lite ? 'md:nx-hidden' : ''}>
+                {mounted ? theme : 'light'}
+              </span>
+            </div>
+          )
+        }}
+      />
+    </div>
   )
 }