+++ packages/nextra-theme-docs/src/constants.tsx
@@ -14,19 +14,17 @@ export const IS_BROWSER = typeof window !== 'undefined'
 
 export const DEFAULT_THEME: DocsThemeConfig = {
   banner: {
-    key: 'nextra-banner',
-    text: ''
+    dismissible: true,
+    key: 'nextra-banner'
   },
   chat: {
     icon: (
       <>
         <DiscordIcon />
-        <span className="sr-only">Discord</span>
+        <span className="nx-sr-only">Discord</span>
       </>
-    ),
-    link: ''
+    )
   },
-  components: {},
   darkMode: true,
   direction: 'ltr',
   docsRepositoryBase: 'https://github.com/shuding/nextra',
@@ -44,15 +42,12 @@ export const DEFAULT_THEME: DocsThemeConfig = {
     },
     text: 'Edit this page'
   },
-  faviconGlyph: '',
-  feedback: {
-    content: null,
-    labels: ''
-  },
+  feedback: {},
   footer: {
     component: Footer,
     text: `MIT ${new Date().getFullYear()} © Nextra.`
   },
+  getNextSeoProps: () => ({ titleTemplate: '%s – Nextra' }),
   gitTimestamp({ timestamp }) {
     const { locale = DEFAULT_LOCALE } = useRouter()
     return (
@@ -81,15 +76,13 @@ export const DEFAULT_THEME: DocsThemeConfig = {
   i18n: [],
   logo: (
     <>
-      <span className="mr-2 hidden font-extrabold md:inline">Nextra</span>
-      <span className="hidden font-normal text-gray-600 md:inline">
+      <span className="font-extrabold">Nextra</span>
+      <span className="ml-2 hidden font-normal text-gray-600 md:inline">
         The Next Docs Builder
       </span>
     </>
   ),
-  main: {
-    extraContent: null
-  },
+  logoLink: true,
   navbar: Navbar,
   navigation: {
     next: true,
@@ -111,11 +104,9 @@ export const DEFAULT_THEME: DocsThemeConfig = {
     icon: (
       <>
         <GitHubIcon />
-        <span className="sr-only">GitHub</span>
+        <span className="nx-sr-only">GitHub</span>
       </>
-    ),
-    // by default should be empty so clicking on project link will go to the github link
-    link: ''
+    )
   },
   search: {
     component({ className, directories }) {
@@ -133,8 +124,10 @@ export const DEFAULT_THEME: DocsThemeConfig = {
     ),
     placeholder() {
       const { locale } = useRouter()
-      if (locale === 'zh-CN') return '搜索文档...'
-      return 'Search documentation...'
+      if (locale === 'zh-CN') return '搜索文档…'
+      if (locale === 'ru-RU') return 'Поиск документации…'
+      if (locale === 'fr-FR') return 'Rechercher de la documentation…'
+      return 'Search documentation…'
     }
   },
   serverSideError: {
@@ -145,10 +138,8 @@ export const DEFAULT_THEME: DocsThemeConfig = {
     defaultMenuCollapseLevel: 2,
     titleComponent: ({ title }) => <>{title}</>
   },
-  titleSuffix: ' – Nextra',
   toc: {
     component: TOC,
-    extraContent: null,
     float: true,
     title: 'On This Page'
   }
@@ -169,7 +160,7 @@ export const DEEP_OBJECT_KEYS = Object.entries(DEFAULT_THEME)
 
 export const LEGACY_CONFIG_OPTIONS: Record<string, string> = {
   bannerKey: 'banner.key',
-  bodyExtraContent: 'main.extraContent',
+  bodyExtraContent: 'main',
   customSearch: 'search.component',
   defaultMenuCollapsed: 'sidebar.defaultMenuCollapseLevel',
   feedbackLabels: 'feedback.labels',
@@ -197,6 +188,7 @@ export const LEGACY_CONFIG_OPTIONS: Record<string, string> = {
 
 export const DEFAULT_PAGE_THEME: PageTheme = {
   breadcrumb: true,
+  collapsed: false,
   footer: true,
   layout: 'default',
   navbar: true,