+++ packages/nextra-theme-docs/src/types.ts
@@ -2,6 +2,7 @@
 import { FC, ReactNode } from 'react'
 import { ThemeProviderProps } from 'next-themes/dist/types'
 import { PageOpts } from 'nextra'
+import { NextSeoProps } from 'next-seo'
 import { Item } from './utils'
 import { TOCProps } from './components/toc'
 import { NavBarProps } from './components/navbar'
@@ -18,14 +19,15 @@ export type RecursivePartial<T> = {
 
 export interface DocsThemeConfig {
   banner: {
+    dismissible: boolean
     key: string
-    text: ReactNode | FC
+    text?: ReactNode | FC
   }
   chat: {
     icon: ReactNode | FC
-    link: string
+    link?: string
   }
-  components: Record<string, FC>
+  components?: Record<string, FC>
   darkMode: boolean
   direction: 'ltr' | 'rtl'
   docsRepositoryBase: string
@@ -37,22 +39,22 @@ export interface DocsThemeConfig {
     }>
     text: ReactNode | FC
   }
-  faviconGlyph: string
+  faviconGlyph?: string
   feedback: {
-    content: ReactNode | FC
-    labels: string
+    content?: ReactNode | FC
+    labels?: string
   }
   footer: {
     component: ReactNode | FC<{ menu: boolean }>
     text: ReactNode | FC
   }
+  getNextSeoProps?: () => NextSeoProps
   gitTimestamp: ReactNode | FC<{ timestamp: Date }>
   head: ReactNode | FC
   i18n: { direction?: string; locale: string; text: string }[]
   logo: ReactNode | FC
-  main: {
-    extraContent: ReactNode | FC
-  }
+  logoLink?: boolean | string
+  main?: FC<{ children: ReactNode }>
   navbar: ReactNode | FC<NavBarProps>
   navigation:
     | boolean
@@ -65,7 +67,7 @@ export interface DocsThemeConfig {
     'defaultTheme' | 'storageKey' | 'forcedTheme'
   >
   notFound: {
-    content: ReactNode | FC,
+    content: ReactNode | FC
     labels: string
   }
   primaryHue:
@@ -76,7 +78,7 @@ export interface DocsThemeConfig {
       }
   project: {
     icon: ReactNode | FC
-    link: string
+    link?: string
   }
   search: {
     component:
@@ -90,18 +92,16 @@ export interface DocsThemeConfig {
     placeholder: string | (() => string)
   }
   serverSideError: {
-    content: ReactNode | FC,
+    content: ReactNode | FC
     labels: string
   }
   sidebar: {
     defaultMenuCollapseLevel: number,
     titleComponent: ReactNode | FC<{ title: string; type: string }>
   }
-  // Can't be React component, otherwise will get Warning: A title element received an array with more than 1 element as children.
-  titleSuffix: string | (() => string)
   toc: {
     component: ReactNode | FC<TOCProps>
-    extraContent: ReactNode | FC
+    extraContent?: ReactNode | FC
     float: boolean
     title: ReactNode | FC
   }
@@ -109,6 +109,7 @@ export interface DocsThemeConfig {
 
 export type PageTheme = {
   breadcrumb: boolean
+  collapsed: boolean
   footer: boolean
   layout: 'default' | 'full' | 'raw'
   navbar: boolean