+++ packages/nextra-theme-docs/src/utils/get-git-issue-url.ts
@@ -1,4 +1,4 @@
-import parseGitUrl from 'parse-git-url'
+import gitUrlParse from 'git-url-parse'
 
 export const getGitIssueUrl = ({
   repository = '',
@@ -9,19 +9,18 @@ export const getGitIssueUrl = ({
   title: string
   labels?: string
 }): string => {
-  const repo = parseGitUrl(repository)
+  const repo = gitUrlParse(repository)
   if (!repo) throw new Error('Invalid `docsRepositoryBase` URL!')
 
-  switch (repo.type) {
-    case 'github':
-      return `https://github.com/${repo.owner}/${
-        repo.name
-      }/issues/new?title=${encodeURIComponent(title)}&labels=${labels || ''}`
-    case 'gitlab':
-      return `https://gitlab.com/${repo.owner}/${
-        repo.name
-      }/-/issues/new?issue[title]=${encodeURIComponent(title)}`
+  if (repo.resource.includes('gitlab')) {
+    return `${repo.protocol}://${repo.resource}/${repo.owner}/${
+      repo.name
+    }/-/issues/new?issue[title]=${encodeURIComponent(title)}`
+  } else if (repo.resource.includes('github')) {
+    return `${repo.protocol}://${repo.resource}/${repo.owner}/${
+      repo.name
+    }/issues/new?title=${encodeURIComponent(title)}&labels=${labels || ''}`
+  } else {
+    return '#'
   }
-
-  return '#'
 }