+++ packages/nextra/src/loader.ts
@@ -72,7 +72,6 @@ async function loader(
     mdxOptions,
     pageMapCache,
     newNextLinkBehavior,
-    allowFutureImage
   } = context.getOptions()
 
   context.cacheable(true)
@@ -119,12 +118,15 @@ async function loader(
     await compileMdx(
       content,
       {
-        mdxOptions,
+        mdxOptions: {
+          ...mdxOptions,
+          jsx: true,
+          outputFormat: 'program',
+        },
         unstable_readingTime,
         unstable_defaultShowCopyCode,
         unstable_staticImage,
         unstable_flexsearch,
-        allowFutureImage
       },
       mdxPath
     )