+++ packages/nextra/CHANGELOG.md
@@ -1,5 +1,46 @@
 # nextra
 
+## 2.0.0-beta.38
+
+## 2.0.0-beta.37
+
+## 2.0.0-beta.36
+
+### Patch Changes
+
+- 1c3fedb: add missing `nx-` prefixes to table/th/tr elements
+
+## 2.0.0-beta.35
+
+## 2.0.0-beta.34
+
+## 2.0.0-beta.33
+
+### Patch Changes
+
+- 580c433: add nx- to all tailwind classes for style isolation
+- 094fdec: sort `defaultMeta` by `frontMatter.date`, if missing by `frontMatter.title` and after by capitalized page name
+- 094fdec: capitalize sidebar's folders names if item is missing in `_meta.json`
+
+## 2.0.0-beta.32
+
+### Patch Changes
+
+- fc8cca0: add `<InformationCircleIcon />` icon, improve `<Callout />` default emojis
+
+## 2.0.0-beta.31
+
+### Patch Changes
+
+- cef5546: allow headings contain links
+
+## 2.0.0-beta.30
+
+### Patch Changes
+
+- 02bc6fc: use `next/future/image` if next>=12.3.0
+- f7856a1: change default options for `compileMdx`, set `jsx: false` and `outputFormat: 'function-body'` by default
+
 ## 2.0.0-beta.29
 
 ## 2.0.0-beta.28